feat(tool):tomatically streams & rules
