+++ tools/kubeedge/README-CN.md
@@ -26,6 +26,14 @@
     "intervalTime":60	//隔多久检查一次命令文件夹（单位：秒）
 }
 
+### 1.4编译程序：
+
+i.在main.go文件所在文件夹下运行命令：go mod init，生产go.mod文件。
+
+ii.将go.mod文件中的“module kuiper/tools/kubeedge”替换为“module kubeedge“。
+
+iii.执行go build命令，生成可执行程序。
+
 ## 2.流的操作示例
 
 ### 2.1.创建流stream1
