--- tools/kubeedge/conf/cf.json
@@ -1,10 +1,0 @@
-{
-    "ip":"127.0.0.1",
-    "port":9081,
-    "logPath":"./log/kubeedge.log",
-    "commandDir":"./sample/",
-    "timeout":500,
-    "intervalTime":60
-}
-
-
