+++ tools/kubeedge/conf/cf.yaml
@@ -1,0 +1,6 @@
+port: 9081
+timeout: 500
+intervalTime: 60
+ip: "127.0.0.1"
+logPath: "./log/kubeedge.log"
+commandDir: "./sample/"
