--- tools/kubeedge/go.mod
@@ -1,8 +1,0 @@
-module kubeedge
-
-go 1.13
-
-require (
-	github.com/benbjohnson/clock v1.0.3
-	github.com/sirupsen/logrus v1.6.0
-)
